SP5: Generic SignalConstSource
