+++ src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.WPF/MainWindow.xaml
@@ -7,22 +7,7 @@
         xmlns:local="clr-namespace:Sandbox.WPF"
         mc:Ignorable="d"
         WindowStartupLocation="CenterScreen"
-        Title="ScottPlot 5 - WPF Sandbox" Height="800" Width="800">
-
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="2*"/>
-            <ColumnDefinition Width="2*"/>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="2*"/>
-            <RowDefinition Height="2*"/>
-        </Grid.RowDefinitions>
-
-        <ScrollViewer Grid.Column="0" Grid.Row="0" Grid.RowSpan="2">
-            <StackPanel x:Name="charts">
-
-            </StackPanel>
-        </ScrollViewer>
-    </Grid>
+        Title="ScottPlot 5 - WPF Sandbox" 
+        Height="600" Width="800">
+    <ScottPlot:WpfPlot Name="WpfPlot1" />
 </Window>
