--- src/ScottPlot5/ScottPlot5/DataSources/SignalConstSource.cs
@@ -1,5 +1,0 @@
-﻿namespace ScottPlot.DataSources;
-
-public class SignalConstSource
-{
-}
