+++ src/ScottPlot5/ScottPlot5/PlottableAdder.cs
@@ -423,9 +423,10 @@ public class PlottableAdder
         return sig;
     }
 
-    public SignalConst SignalConst(double[] ys, double period = 1, Color? color = null)
+    public SignalConst<T> SignalConst<T>(T[] ys, double period = 1, Color? color = null)
+        where T : struct, IComparable
     {
-        SignalConst sig = new(ys, period)
+        SignalConst<T> sig = new(ys, period)
         {
             Color = color ?? GetNextColor()
         };
