+++ src/ScottPlot5/ScottPlot5/Plottables/SignalConst.cs
@@ -2,20 +2,14 @@
 
 namespace ScottPlot.Plottables;
 
-public class SignalConst : IPlottable
+public class SignalConst<T>(T[] ys, double period) : IPlottable
+    where T : struct, IComparable
 {
-    readonly SegmentedTree<double> SegmentedTree = new();
-    readonly double[] Ys;
-    readonly double Period;
-
+    readonly SignalConstSourceArray<T> DataSource = new(ys, period);
     public readonly MarkerStyle Marker = new();
     public readonly LineStyle LineStyle = new();
 
     public string? Label { get; set; }
-    public double XOffset = 0;
-    public double YOffset = 0;
-    public int MinRenderIndex = 0;
-    public int MaxRenderIndex = int.MaxValue;
 
     public Color Color
     {
@@ -28,84 +22,12 @@ public class SignalConst : IPlottable
         }
     }
 
-    public SignalConst(double[] ys, double period = 1)
-    {
-        Ys = ys;
-        Period = period;
-        SegmentedTree.SourceArray = ys;
-        MaxRenderIndex = ys.Length - 1;
-    }
-
     public bool IsVisible { get; set; } = true;
     public IAxes Axes { get; set; } = ScottPlot.Axes.Default;
 
     public IEnumerable<LegendItem> LegendItems => LegendItem.None;
 
-    public AxisLimits GetAxisLimits()
-    {
-        SegmentedTree.MinMaxRangeQuery(0, Ys.Length - 1, out double low, out double high);
-        return new AxisLimits(0, Ys.Length * Period, low, high);
-    }
-
-    public int GetIndex(double x, bool visibleDataOnly)
-    {
-        int i = (int)((x - XOffset) / Period);
-
-        if (visibleDataOnly)
-        {
-            i = Math.Max(i, MinRenderIndex);
-            i = Math.Min(i, MaxRenderIndex);
-        }
-
-        return i;
-    }
-
-    public bool RangeContainsSignal(double xMin, double xMax)
-    {
-        int xMinIndex = GetIndex(xMin, false);
-        int xMaxIndex = GetIndex(xMax, false);
-        return xMaxIndex >= MinRenderIndex && xMinIndex <= MaxRenderIndex;
-    }
-
-    public SignalRangeY GetLimitsY(int firstIndex, int lastIndex)
-    {
-        double min = double.PositiveInfinity;
-        double max = double.NegativeInfinity;
-
-        for (int i = firstIndex; i <= lastIndex; i++)
-        {
-            min = Math.Min(min, Ys[i]);
-            max = Math.Max(max, Ys[i]);
-        }
-
-        return new SignalRangeY(min, max);
-    }
-
-    public PixelColumn GetPixelColumn(IAxes axes, int xPixelIndex)
-    {
-        float xPixel = axes.DataRect.Left + xPixelIndex;
-        double xRangeMin = axes.GetCoordinateX(xPixel);
-        float xUnitsPerPixel = (float)(axes.XAxis.Width / axes.DataRect.Width);
-        double xRangeMax = xRangeMin + xUnitsPerPixel;
-
-        if (RangeContainsSignal(xRangeMin, xRangeMax) == false)
-            return PixelColumn.WithoutData(xPixel);
-
-        // determine column limits horizontally
-        int i1 = GetIndex(xRangeMin, true);
-        int i2 = GetIndex(xRangeMax, true);
-
-        // first and last Y vales for this column
-        float yEnter = axes.GetPixelY(Ys[i1] + YOffset);
-        float yExit = axes.GetPixelY(Ys[i2] + YOffset);
-
-        // column min and max
-        SignalRangeY rangeY = GetLimitsY(i1, i2);
-        float yBottom = axes.GetPixelY(rangeY.Min + YOffset);
-        float yTop = axes.GetPixelY(rangeY.Max + YOffset);
-
-        return new PixelColumn(xPixel, yEnter, yExit, yBottom, yTop);
-    }
+    public AxisLimits GetAxisLimits() => DataSource.GetAxisLimits();
 
     public void Render(RenderPack rp)
     {
@@ -113,7 +35,7 @@ public class SignalConst : IPlottable
         LineStyle.ApplyToPaint(paint);
 
         IEnumerable<PixelColumn> cols = Enumerable.Range(0, (int)Axes.DataRect.Width)
-            .Select(x => GetPixelColumn(Axes, x))
+            .Select(x => DataSource.GetPixelColumn(Axes, x))
             .Where(x => x.HasData);
 
         if (!cols.Any())
